+++ src/ethereum/info/ethereumInfo.ts
@@ -815,11 +815,19 @@ export const builtinTokens: EdgeTokenMap = {
     }
   },
   a4bdb11dc0a2bec88d24a3aa1e6bb17201112ebe: {
+    currencyCode: 'USDSC',
+    displayName: 'StableUSD Classic',
+    denominations: [{ name: 'USDSC', multiplier: '1000000' }],
+    networkLocation: {
+      contractAddress: '0xA4Bdb11dc0a2bEC88d24A3aa1E6Bb17201112eBe'
+    }
+  },
+  dc035d45d973e3ec169d2276ddab16f1e407384f: {
     currencyCode: 'USDS',
     displayName: 'StableUSD',
     denominations: [{ name: 'USDS', multiplier: '1000000' }],
     networkLocation: {
-      contractAddress: '0xA4Bdb11dc0a2bEC88d24A3aa1E6Bb17201112eBe'
+      contractAddress: '0xdC035D45d973E3EC169d2276DDab16f1e407384F'
     }
   },
   dac17f958d2ee523a2206206994597c13d831ec7: {